+++ R/crew_client.R
@@ -45,7 +45,7 @@ crew_client <- function(
   tls_enable = NULL,
   tls_config = NULL,
   serialization = NULL,
-  seconds_interval = 1,
+  seconds_interval = 0.25,
   seconds_timeout = 60,
   retry_tasks = NULL
 ) {